fixes dump rings choosing in autotests
Information about dump rings was previously added to a map with ring tag
as a key. This resulted in ring information being overwritten by next ring
with the same tag, leaving only the last parsed ring information.

This commit switches to using ring unique name as a key, tests are
changed accordingly